Add sim connection option to input args, add basic LED spec (#12)
* Add sim connection option to input args, add basic LED spec

* Make sim connection return unattached devices on discovery

* Indicate via the device address that its a simulation

* Change device discovered identifier to generic 'identifier' instead of 'address'

* Add connect/disconnec to GCI of sim

* Add support for UNPAIR

* update manage.py, remove clean-db opt arg, remove clear-device-data subparser

* Don't recreate existing devices if multiple discoveries, don't simulated device discovered events if no devices were found

* add detach support

* linting

* extend GCI with for_each

* add heartbeat support

* register DeviceHealth model

* first heartbeat on hume start is done quickly, then at 60 second interval, detach leads to health clear

* add attach failure handling and include the device identifier from discovery in capability request

* remove todo

* linting

* Respond to stateful actions to HINT when device responds with what the new state is and if it was a success

* lint

+++ hume/storage/data_store.py
@@ -76,8 +76,20 @@ def delete(obj):
     :param obj:
     :return:
     """
+    if obj is None:
+        LOGGER.warning("tried to delete None")
+        return
+
     LOGGER.info("deleting object")
     _store.delete(obj)
+
+
+def delete_all():
+    """
+    Delete all data.
+    """
+    LOGGER.info("deleting all data")
+    _store.delete_all()
 
 
 class DataStore:
@@ -150,7 +162,7 @@ class DataStore:
     def get(self, cls, key, **kwargs):
         """
         Get a single object matching the provided key. Will always check local
-        storage only as it should be up to date with persistent storage.
+        storage only as it should be up-to-date with persistent storage.
 
         :param cls: class
         :param key: key
@@ -173,8 +185,17 @@ class DataStore:
 
         :param obj:
         """
-        self._persistent_storage.delete(obj)
+        if issubclass(obj.__class__, PersistentModel):
+            self._persistent_storage.delete(obj)
+
         self._local_storage.delete(obj)
+
+    def delete_all(self):
+        """
+        Clears data from all registered tables.
+        """
+        self._local_storage.delete_all()
+        self._persistent_storage.delete_all()
 
 
 _store = DataStore()
